refactor: moved tests constants to class and changed imports to static in integration tests
