+++ am-lib/src/test/java/integration/uk/gov/hmcts/reform/amlib/GetAccessorsListIntegrationTest.java
@@ -6,9 +6,10 @@ import org.junit.Test;
 import java.util.List;
 import java.util.UUID;
 
-import static integration.uk.gov.hmcts.reform.amlib.Constants.ACCESSOR_ID;
-import static integration.uk.gov.hmcts.reform.amlib.Constants.EXPLICIT_READ_CREATE_UPDATE_PERMISSIONS;
-import static integration.uk.gov.hmcts.reform.amlib.Constants.OTHER_ACCESSOR_ID;
+import static integration.uk.gov.hmcts.reform.amlib.TestConstants.ACCESSOR_ID;
+import static integration.uk.gov.hmcts.reform.amlib.TestConstants.EXPLICIT_READ_CREATE_UPDATE_PERMISSIONS;
+import static integration.uk.gov.hmcts.reform.amlib.TestConstants.OTHER_ACCESSOR_ID;
+import static integration.uk.gov.hmcts.reform.amlib.TestConstants.createRecord;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class GetAccessorsListIntegrationTest extends IntegrationBaseTest {
