+++ am-lib/src/test/java/integration/uk/gov/hmcts/reform/amlib/GrantAccessIntegrationTest.java
@@ -5,8 +5,10 @@ import org.junit.Test;
 
 import java.util.UUID;
 
-import static integration.uk.gov.hmcts.reform.amlib.Constants.ACCESSOR_ID;
-import static integration.uk.gov.hmcts.reform.amlib.Constants.EXPLICIT_READ_CREATE_UPDATE_PERMISSIONS;
+import static integration.uk.gov.hmcts.reform.amlib.TestConstants.ACCESSOR_ID;
+import static integration.uk.gov.hmcts.reform.amlib.TestConstants.EXPLICIT_READ_CREATE_UPDATE_PERMISSIONS;
+import static integration.uk.gov.hmcts.reform.amlib.TestConstants.countResourcesById;
+import static integration.uk.gov.hmcts.reform.amlib.TestConstants.createRecord;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class GrantAccessIntegrationTest extends IntegrationBaseTest {
